+++ src/config/periods.js
@@ -1,10 +1,18 @@
-import endOfDay from 'date-fns/endOfDay';
+import {
+  endOfDay,
+  getQuarter,
+  subQuarters,
+  startOfQuarter,
+  endOfQuarter,
+} from 'date-fns';
 import i18n from '@/locale/i18n';
 
 export default function getPeriods() {
   const currentDate = new Date();
   const endDate = endOfDay(currentDate);
   const currentYear = currentDate.getFullYear();
+  const currentQuarter = getQuarter(currentDate);
+  const prevQuarterDate = subQuarters(currentDate, 1);
   const sixMonthsBack = new Date();
   sixMonthsBack.setMonth(sixMonthsBack.getMonth() - 6);
 
@@ -26,6 +34,20 @@ export default function getPeriods() {
       key: 'sixmonths',
       startDate: sixMonthsBack,
       endDate,
+    },
+    quarter: {
+      label: `${i18n.t('period.thisQuarter')} (Q${currentQuarter} ${currentYear})`,
+      key: 'quarter',
+      startDate: startOfQuarter(currentDate),
+      endDate: endOfQuarter(currentDate),
+    },
+    prevquarter: {
+      label: `${i18n.t('period.previousQuarter')} (Q${getQuarter(
+        prevQuarterDate
+      )} ${prevQuarterDate.getFullYear()})`,
+      key: 'prevquarter',
+      startDate: startOfQuarter(prevQuarterDate),
+      endDate: endOfQuarter(prevQuarterDate),
     },
   };
 }